+++ ACadSharp/Entities/PolyLine2D.cs
@@ -1,5 +1,8 @@
 ﻿using ACadSharp.Attributes;
+using CSMath;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace ACadSharp.Entities
 {
